refactoring: Record and Deposit resolvers
